Handle multiple SMTP messages with one connection.
mailcatcher has had a long-standing issue whereby it doesn't handle
multiple mails over the same connection.  This is actually a problem
with the SMTP server in EventMachine itself, but the fix still hasn't
been merged into EM yet.  There's an open pull request, and the
easiest way is to monkey-patch what we need.  For more details:
https://github.com/railsware/eventmachine/commit/a23d4828